refactor(input): split parser.rs by command family (TD-030)
Behavior-preserving split of parish-input/src/parser.rs (~1216 LOC) into
five submodules by command family:

- parser/save.rs    — /fork, /load
- parser/world.rs   — /map, /wait, /unexplored, /weather
- parser/display.rs — /theme, /debug, /spinner, /speed
- parser/inference.rs — /preset, /provider, /model, /key, /cloud,
                        /inference-log, dot-notation category variants
- parser/flags.rs   — /flag

parser.rs is reduced to a hub file that declares the submodules and
re-exports the public API unchanged (parse_system_command, classify_input).
Tests moved alongside the code they exercise. All 3089 workspace tests
pass; architecture_fitness (5 tests) passes; clippy and fmt clean.

Part of #1200 (TD-030).
